minor bugfixes & improvements
Changes made:
  * autonomous.zig:
    * (~) updated to use the rotateDeg func instead of rotate
    * (#) fixed duplicate driveRight (instead of driveLeft) calls
    * (#) optimized the wait function predicate
  * pid.zig:
    * (#) added a missing @abs to the move func's distance calculation
    * (~) changed rotate to rotateDeg so I stop making errors

The main explanation, is that rotate is a convenience function.
It's whole purpose is to make the blocking state-machine trivial to
implement, and be as concise as possible (no matter how correct it may
be). Anyhow, I would never use this function and want to type in
radians. So I've just baked in degrees into the function, and renamed
it to remove any remaining ambiguity.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>