physical changes accomodated
Changes made:
  * tower.zig:
    * (~) added extra middle tower motor
    * (~) added support for tower motor gearset variation
    * (~) added motor tower controls (3 motors)
    * (~) updated flake.lock (update)

Recently the robot tower structure changed once more so that now it has
three motors: the top, middle and bottom.

There are to be three modes of usage:
  1. scoring - all spinning up (R1 & R2)
  2. out-take - all spinning down (R1)
  3. storing - bottom two spinning up, top braking (R2)

Storing is default as we wish to be able to collect multiple blocks
without accidentally scoring outside of a goal.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>