simulation: initialized basic build step & stubs
Changes made:
  * (+) build.zig: added the simulation build step and all the linking logic
  * (+) simulation.zig: added a basic entry point and 'routine'.
  * (+) simulation/stubs.zig: added basic stubs with a debug printf 'implementation'

I've been working on a non-physical Vex brain simulator.

The basic idea is this:

How the entire program works, is that it links to the pros library to do
all the IO and actions on the vex robot brain.
Also all the entry points into the program are also clearly defined for
PROS which makes everything incredibly simple.

So simulation is as simple as linking the program to a custom simulation
that mimics the pros library's behaviour on a real robot brain.

So as long as I implement the functions and have some sort of state
machine for the simulated robot (and kinematics) then I can test out
the code that will actually be running and not just a simple unit test
scenario.

To test out the idea so far I've just added a printf and millis
implementation and so far, everything is working exactly as I'd expected.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>