added simulated io, formatting and fs
Changes made:
  * (~) .gitignore: don't include simulation-fs/ (files generated by the sim)
  * (~) flake.lock: updated zig compiler
  * (+) main.zig: added some io, formatting and fs testing lines (remove later)
  * (+) simulation.zig: added a global managed allocator for the simulation
  * (~) stub.zig: made any unimplemented stubs crash instead of silently fail
  * simulation/io.zig:
    * (+) added path sandboxing (change the `/usd/`)
    * (+) added a custom c-style string formatting function
    * (+) added custom implementations of the printf, fopen, fprintf, and fclose functions

Everything should be self-explanitory.
All paths in pros start with the prefix '/usd/'.
So if we just remove it and then use the the cwd simluation-fs, we can
sandbox the fs.

I've made it crash for unimplemented things instead of silently fail and
output undefined memory garbage for obvious reasons.

Implementing formatting was fun with a custom quick-ly written parser.

The main 'hacky' thing that actually deserves explaination is how the
*FILE pointer is handled.
Essentially, in zig, all the FILE type is, is an opaque type.
That just means that it's a type that the language knows nothing about:
no size, internal or anything.
We can only interact with it through a pointer; a pointer that is created
and 'closed' by the fopen and fclose functions.
Functions that we implement.
So, the pointer we return doesn't actually have to be the C *FILE and
can be instead whatever we choose; there's no difference to zig.
And it's fine since we (the simulation) manages all the memory allocation
and etc, and we return the correct kind of type (a pointer).
That's kind of the whole point of the opaque type.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>