tower: extra physical hood motor
Changes made:
  * (-) logging.zig: removed velocity and temperature logs
  * tower.zig:
    * (~) added hood motor
    * (~) inverted and switched to 5.5W for bottom motor
    * (~) set storing speed to be max

Recently there's been another change to the robot tower:
there's been the addition of a hood motor.
The behaviour of the hood motor should be the exact same as the top roller.
Also, I've changed it to make the hood roller spin backwards instead of brake as that seems to fit more blocks into the tower.

Logging for the velocity and motor temperature has been removed.
As logging as a whole is basically useless as I have never needed to use it aside from a gimick to impress people.
And it's been a real pain whenever I've had to modify one of the tower motors.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>