driver feedback changes
Changes made:
  * drive.zig
    * (~) switched reverse toggle to be the LEFT button
    * (~) changed driving and turning to be full speed
    * (~) switched from velocity to volts (too exponential)
  * (#) odom.zig: quickfix to nullify lateral odom (set offset to 0)
  * (~) tower.zig: boosted tower outtake speeds to max

After testing at the AI Kings Boxhill scrim, these are the changes we've
made to improve the driving experience.

I've changed the reverse toggle to be the left button instead of down
as it's never used and down can be hit easily with controller mods and
has therefore been reserved for some cooler functionality.

Also, it's a bit of a waste to have a fast drive-train and not use the
speed.

Recently the driver, and everyone for that matter, have been complaining
that there were giant dead-zones and that driving was too 'exponential'.
In that, when holding a speed, the robot would speed up slowly instead
of instantly.
This turned out to be due to me using velocity (mainly due to the I term)
instead of voltage (which I am now doing).

Also, after brief testing in the simulation, it was discovered that
even with a zero value from the lateral sensor, it still messed with
odom values, unless the offset was set to zero.
We don't have a lateral odom sensor anymore.

Also, tower outtake is too slow now, so I've had to boost it in code.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>