removed redundant code & changed drive & tower code
Changes made:
  * (-) build.zig: removed now redundant/deprecated options
  * drive.zig:
    * (+) added a cleaner centralized state for drive code
    * (+) added an 'absolute turning' drive mode (similar to actual car driving) binded to controller mod down
    * (-) removed redundant split arcade and tank drive options
  * tower.zig:
    * (+) added centralized state for tower logic
    * (~) made intake toggle instead of hold
    * (~) made little will toggle instead of two buttons

Based upon driver feedback and a few ideas I've had, I've 'improved' the
drive code once more.

I've also removed a bunch of redundant code; code that doesn't make sense
to have since we know we'll never need that option.
Such as tank drive -- we'll never need to use tank drive or normal
arcade drive so those options are pointless.

Toggle arcade has proven to be useless and now has been removed -- with
split arcade now being the default.

I've also centralized state for both drive and tower logic as it's a real
pain having to pass in all the state variables one by one.

I've also implemented an experimental driving controls that I thought
would help with aligment -- 'absolute turning'.
In that, when the controller-mod down key is held, the right joystick
will instead dictate the absolute angle that the robot will face
(through the use of a correctly tuned pid).
And you can use the left joystick as a 'thrust' -- similar to how
actual cars drive (kinematic bicycle model).

The driver has also requested that both little-will and the robot intake
be toggle-able instead of held down as it's easier to remember the
controls that way.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>