experimental PID debug mode
Changes made:
  * (~) build.zig: added pid debug mode
  * (#) debug_mode.zig: added the required conversion to volts
  * (+) debug_mode_pid.zig: cloned debug_mode but made it for tuning movement & rotational PIDs
  * (~) pid.zig: changed the move function to not use coordinates

I'd wanted something similar to debug_mode's live parameter changing and
tuning.

So, instead of having to re-implement all of that from scratch, I've
chosen instead to just copy paste debug_mode and eye-ball tailor it
to tune the yaw & movement PIDs.

Also, since I'm planning to use the PID move blocking-func as like
the one in python -- where it just moves a distance.
That's the function signature that I'd like as that's the easiest to
just hand calculate and use.

But doesn't it use a path and coordinates?
How would you make that work with just a pid?
Well, it's pretty simple, I'll just have it follow the path by first
rotating towards the point, and then moving towards it, and repeating
until all the path points are reached.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>