GIANT COMMIT: FULLY WORKING KINEMATICS & DEVICE SIM
Changes made:
  * (~) autonomous.zig: switched to use pp instead of autonNew routine (for testing)
  * (#) logging.zig: added missing % escape in string format
  * simulation.zig:
    * (+) added global simulated kinematics state
    * (~) removed opcontrol from routine (no simulated controller yet)
  * simulation/State.zig:
    * (+) added struct for storing the state of the simulated robot
    * (+) added forwards kinematics simulation
  * (+) simulation/configs.zig: added global simulation configurations
  * simulation/devices.zig:
    * (+) simulated API binding `imu_reset`
    * (+) simulated API binding `motor_set_encoder_units`
    * (+) simulated API binding `motor_set_gearing`
    * (+) simulated API binding `adi_port_set_config`
    * (+) simulated API binding `rotation_set_reversed`
    * (+) simulated API binding `rotation_get_angle`
    * (+) simulated API binding `motor_move_velocity`
    * (+) simulated API binding `millis`
    * (+) simulated API binding `task_delay_until`
  * (+) simulation/io.zig: added support for formatting unsigned integers

 I don't want to explain and go over everything that's been written
 because there is A LOT.

 I'll only explain what needs explaining.

First, the reason why things like reversing motors or roation sensors
are just ignored, is that our simulation isn't aiming to be physically
correct. So yes, physically, the right side should be reversed and then
the function call would undo that -- but we could just simplify it out
and have the motors and rotation sensors not need reversing from the
get-go.

The next thing that needs real explaining is how state works and how
it's updated in this simulation.
So anything that happens in the program needs time for it to have any
effect.
Using that to our advantage, we could have all the kinematics simulations
be done when the wait function is called.

"Oh but the function calls that are done before he wait would get
ignored!"
Yes but that would happen in real life anyways. As, if you set the speed
of a motor to a value and then change it within the micro-second, it's
not even going to be visible physically.

This makes our life even easier as we are also supplied the delta time
through this and can simulate faster than real-time.

So, all the device calls just modify the state of the simulation, and
once the program calls the wait function, the modified state comes
into effect and is used in the kinematics calculations to find the
next yaw and coordinate etc.

Also, any motor movements are also stored in an array of 21 as I'm lazy
but still want to have the ability to graph them if I so please.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>