simulation: added logging
Changes made:
  * (+) simulation.zig: added a buffered global log file
  * simulation/State.zig
    * (#) forward kinematics used seconds instead of milliseconds
    * (+) added a log function
  * (+) simulation/devices.zig: added log to task_delay

I've added logging of all the simulation state so we can actually
debug.

The one bug that I did discover was that the simulated robot was going
WAYY to slow. This turned out to be due to me accidentally using
seconds instead of milliseconds in the forwards kinematics function.

Signed-off-by: Ethan Ma <kalscium@protonmaill.com>