shadow: added imperative commands to pure pursuit
Changes made:
  * (+) Shadow.zig: added an imperative shadow for pure pursuit to follow
  * (+) autonomous.zig: added some shadow showcase code
  * (#) pure_pursuit.zig: fixed reversed driving issues by using odom_copy instead of odom_state

The main issue hindering writing autonomous, is the pain that is getting
and writing coordinates.
So far, my solution to that problem was to drive the robot to the points,
and 'record' the coordinates before copying them.
I'm no longer doing that as I find that extremely fustrating, esoteric
and it just doesn't work.

Instead I've had the idea to somehow use imperative commands like
`move_forwards` or `turn` and convert them into coordinates.
How I plan to do this is to use a 'shadow' in which all these
transformations happen, and then I just save the coordinates for pure
pursuit.

Also, for some reason, I wasn't using the odom_copy where all the
transformations to get reversed driving to work, so I've just changed
that.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>