making dual pursuit actually viable (yaw accuracy)
Changes made:
  * (~) autonomous.zig: updated and eyeballed some new parameters for pure pursuit
  * pure_pursuit.zig:
    * (#) used the search radius instead of the bounding radius on occasion
    * (+) added an adjustable yaw-limit to tune the accuracy of dual pursuit (instead of just 90)
    * (+) added a speed controller that also slows at the start of a path
    * (~) simplified a bunch of parameter comments

The LARGEST issue with pure or dual pursuit is the GIANT errors that it
has whilst turning. And yes, that is sort of the whole point of pure
pursuit -- having smooth turns, but to have driving that is smooth
the turns made are unacceptably large.
The solution to this I've found, is to just to have a smaller limit
to when the arc is fitted, and when it is just rotated towards it.
I've made this a parameter so it's easier to tune.

Also, with the speed controller.
So far how it works, is that until the end of the path segment is reached
the goal point will always be the search radius away from the bot; so
the speed would be 1. Once the path segment end is within the search
radius however, the distance shrinks and so does the speed.
The speed never reaches 0 however, as it crosses the bounding-radius and
the path_segment_end is incremented to the next end (out of the radius).

This usually doesn't work for the start of a path segment, due to nothing
stopping the speed from reaching zero and making the robot stop.
Unless we just put a clamp there.
If we combine the two with another comparison, then you can have a
continuous slowdown within the search radius around the path end.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>