WORKING AUTON 2!!!!!!!!!!!!!!
Changes made:
  * (+) build.zig: added w-akibot flag (wait 2s in auton to not collide)
  * (+) Shadow.zig: added rotate and movement functions with PID
  * (+) autonomous.zig: added a new auton-left routine that starts at the front of the parking zone
  * (+) pid.zig: added a function to move to a coord (no rotation) with a PID
  * (+) vector.zig: added a vector normalize func I don't use

After yet another auton testing session, we've made some significant
advancements.

So the main code advancement is the usage of the Shadow function with
PID movement. So it's essentially the same as before -- using the PID
functions -- but it also keeps track of the robot's 'perfect position'
so that any inacuracies in movement don't propagate
(only yaw errors matter now).
Also the shadow uses the pid move to coord func with the 'ideal' coord.

Now that I think about it, I could also have the movement pid use
circle fitting to the entire end point as there shouldn't be a large
yaw error ever and that'll correct all the issues.

Now, the main advancement in the auton routine is that we've decided
that, since the robot at the start rotates AGAINST the parking it's
final lateral position is incredibly inaccurate.
So we've moved the robot to be in front of the parking aligned to the
middle line.
But this is touching ehen's robot's auton, so we've also included a
build switch now to wait 2s to not colide with them (only when alianced)
with them.

Signed-off-by: Ethan Ma