follow arc shadow pid move (revert if not working)
Changes made:
  * (~) pid.zig: made the move coord function use follow arc to correct minor yaw errors

I've added a follow arc function into the move coord func of the pid
(that get's called by the shadow) so it should correct any minor yaw
errors now.

It follows the arc from the start to the end of the path
(no search radius) so the error in yaw should be extremely low as the
turning/rotation is done before the move so any yaw errors should be
within a degree or two.

This is largely to just combat any accumulative large drifts from minor
yaw errors.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>