competition commit (procrastination date)
Changes made:
  * autonomous.zig:
       * (~) tweaked auton left values at comp to work better most of the time
       * (~) made auton right to just be auton left with flipped yaw
  * (#) pid.zig: removed circle fitting as it causes the robot to tweak
  * (#) tower.zig: made intake toggle off when scoring or outtaking

On saturday we had our competition at CopperField.

The first change made was -- after some brief auton-testing -- to remove
circle fitting from the pid move coord function.
The reason was that it would overcorrect/turn and the spinning circle
would knock the blocks out of the way whilst getting to them.

I've tweaked the auton values to work better during comp whilst observing
the games and it's in a really good state (BEFORE SOMETHING).

Right auton is just left auton but flipped yaw currently
(it doesn't work tho).

Oh yeah, the reason why auton worked flawlessly on quarter-finals but
not semi-finals, was not due to random error -- but rather a last minute
change I made before uploading.
The change? I wanted the robot to travel faster whilst intaking the
blocks so I changed the speed from 0.1 to 0.15.
This is fine on it's own, but I forgot to do it for both sides, so it
veered off course and messed up all the alignments.

(This will be fixed later by having the robot just rotate towards the
shadow coord).

Also, our driver had complained that due to the toggle, after they
finished out-taking or scoring, the tower would start intaking in as they
never explicitly toggled it off.
This behaviour is now changed to be more intuitive so that any other
action overrides it and toggles it off.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>