no reverse drive & auton pid moveCoord improvements
Changes made:
  * (-) drive.zig: removed reverse drive toggle
  * (~) pid.zig: added an AOT yaw correction before moving

I've removed the driving in reverse toggle as our driver has NEVER used
it and has never *needed* to use it.
I've explicitly removed it as, if our driver were to accidentally hit it,
then it would reverse and mess up everything.

Also, to correct accumalative errors from slight yaw miss-alignments.
The main issue with tinkering with pure pursuit and the such, I've
discovered that live yaw-correction whilst moving/driving is just not
viable as it's FAR too noisy.
Instead, a better approach is to just do it AOT (ahead of time), where
before moving (or in between movements), you correct the yaw or calculate
the necessary angular velocity to get to that goal point and then do it
until the closest point to the goal is reached.
So the robot keeps a constant angular velocity whilst moving.
It's like taking a look at your surrounding, and then closing your eyes
and running 'til you hit something.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>