experimental held opcontrol yaw when not turning
Changes made:
  * drive.zig:
    * (-) removed userArcadeDrive by merging functionality with normal arcadeDrive
    * (+) added holding of last turned yaw when not turning
    * (~) simplified controller drive function code

I've removed userArcadeDrive by merging it's functionality with the
normal arcade drive and shifting the logic to the actual user drive
function.

I've also implemented holding of the last turned yaw when the robot is
not turning (x == 0) through the yaw PID (that's also used for abs turn).
This also has the side effect of the robot using velocity for driving
when not turning which might either be beneficial or incredibly annoying.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>