another auton testing sess -- MAGESTIC SKILLS PARK
Changes made:
  * autonomous.zig:
    * (~) tweaked the auton values to intentionally curve to better intake blocks
    * (~) further tuned left auton values to score really well
    * (+) added parking at the end of left auton (rudimentary but works 3/3 times and looks cool)
    * (+) added a right auton that's just the left but mirrored
  * (-) drive.zig: removed maintained last turned yaw
  * (~) pid.zig: added a little yaw correction at the start of moveCoord

Main notes.

We've realized that the curving whilst intaking blocks is actually really
beneficial and so I've added a little bit of code that makes the robot
point 'towards' the goal with the side it's going to drive with before
driving towards it.
It takes a little bit of time, so perhaps circle fitting might be better
suited for this.

Also, the maintained yaw driving is incredibly jank and we've decided
to just discard it.

I've also implemented a simple parking for skills and it WORKS SO WELL.
It just rams into the goal but due to hitting it with half the robot,
the robot curves and goes into the goal 3/3 times we tested it.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>