simplified auton picking
Changes made:
  * autonomous.zig:
    * (~) simplified auton picking to be one big routine with unified state (allows for chaining & extensions)
    * (+) moved parking into it's own skills auton routine
  * (#) zig.mk: added itself to it's dependencies so that changes in routine actually rebuild

I've simplified auton by having each routine instead of being it's own
routine with independant state -- just be actions that can be applied
to a global autonomous state.
This allows for easy chaining of auton routines and also extensions of
pre-existing routines.
And this is what I've done with the skills routine (just a copy of left
+ parking).

Also I've added zig.mk to it's own dependency list as before, if I made
a change to the command to switch auton routines, it wouldn't rebuild
the user binary and I'd have to `rm -rf bin` to get it to work.
And worse if I forget, then that would have the wrong routine.
It's fixed now.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>