trueFollowArc moveCoord correction & motion chains
Changes made:
  * (~) Shadow.zig: added a little note to try motion chaining
  * (+) autonomous.zig: added a speed config for motion chain driving
  * pid.zig:
    * (~) replaced moveCoord rotation correction with circle fitting
    * (+) added a motion chain option for driving

I keep wanting to have at least a little bit of the work I put into pure
pursuit be of use, so I've replaced the old way of correcting yaw errors
before moving with a circle-fit instead of a rotation.

It follows my philosophy of separating the movement and the correction
of yaw (as that's incredibly noisy).
And, since it's for tiny yaw inconsistencies, it should be really smooth.

I've also looked into motion-chaining and implemented my own rendition
of it that fits in our code.
The basic idea is to just cut out the PID and use a constant speed for
movement that stops once it surpasses the reachable goal point (when
the dot-product distance flips signs).
And then fix any errors afterwords (through circle fitting of the next
go).

It's not enabled, but can be for testing.

"Just do it first -- you can fix it later."

Signed-off-by: Ethan Ma <kalscium@protonmail.com>