yet another physical tower change
Changes made:
  * (+) tower.zig: removed top and added storage motor
  * (+) logging.zig: removed top and added storage motor

Recently I've been informed by the builder that the top motor of the
tower has just been combined with the middle tower motor and a 'storage'
motor has been added that moves the blocks into the storage section of
the bot.

We shall add color sort soon, perhaps manually for now.

Signed-off-by: Ethan Ma <kalscium>