updated to follow new robot & driver changes
Changes made:
  * (+) build.zig: added a practice-driving mode
  * (-) autonomous.bck: removed a random backup file
  * drive.zig:
    * (~) updated drivetrain ports
    * (~) replaced old log controls with daniel's magic equation
    * (+) added a practice mode for driving
  * tower.zig:
    * (~) changed tower triggers to combat stick
    * (#) rewrote the logic for the tower to be more cohesive
    * (+) added a second controller storage discard control

This robotics session introduced a bunch more changes to the tower.
So now there was a primative storage system/motor, that upon the press of
a button on either the master or partner controller would store the block
into the larger storage container.
This acts as a more primative color-sort as we would have the partner
driver look out for blocks of the wrong color and then hold the button
to reverse the 'storage' motor.

Also, after doing some research online, I've discovered a thing called
'driver practice modes'.
Where, to train drivers not to just go from 0-1 and 1-0, they made 1
jump to zero so the drivers would use more of the joystick's range.

Also, daniel has supplied me with yet another magic number -- in the form
of a joystick scaling algorithm that starts off logarithmic and then
ends nice and linear.
This overcomes the dead-zone whilst maintaining an intuitive linear
control.

Signed-off-by: Ethan Ma <kalsicum@protonmail.com>