parking-aligned autonomous
Changes made:
  * (+) autonomous.zig: added new parking-aligned left auton routine

Since the largest issue with our current auton is lining up and speed,
we've decided to redo auton but at a higher speed and also have it
aligned against the wall and the parking zone so that there are no
possible ambiguities as to where it should be.

Whilst planning this route something incredibly unexpected happened --
a neat turn that both intakes the blocks and also pivots the robot into
the middle goal.

We also got match-loading working.
The only issue that we didn't have enough time to fix (we'll do it at
comp) is after match-loading, when the robot spins around and it whacks
the goal with little-will miss-aligning it.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>