added support for pneumatic parking (controls)
Changes made:
  * (+) controller.zig: added support for 'any' controller digital input and more input functions
  * tower.zig:
    * (-) removed 'storage' stuff -- back to top motor.
    * (+) added parking controls with pneumatics etc
  * (~) logging.zig: back to top instead of storage tower motor

A new exciting physical change/feature of the robot has been added.

Since the bot is allowed to rest on blocks and count as 'parked', our
builder has decided to allow our bot to park through jacking itself up
with a block in it's intake.

How it works is that the intake will have a block in it, and a piston
will push on the intake -- pushing the entire bot up.

The controls that I've implemented for it are pretty simple and can be
dissolved into one button really.

The Y button.
When held, it modifies the tower behaviour -- with a much slower speed
to allow for precise positioning of the block.
And, when released, it activates the parking piston and jacks the robot
up.
If there's been a mistake made, then simply drop little-will, and the
parking piston will also dis-engage, as little will would never be down
during parking anyways and that simplifies the controls.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>