Rivercrest competition commit
Changes made:
  * autonomous.zig:
    * (~) improved auton speed by reducing yaw precision
    * (+) added a simple park auton
    * (~) tweaked old auton to make it work on comp field
    * (!) parking doesn't work no more
    * (~) tried to fix parked auton (not working now)
  * tower.zig:
    * (+) added support for parking that instantly got removed

Today was the rivercrest competition (we came first in both qualli and
finals).

Many changes were made on the spot today that I'll outline now.

One main thing we noticed during auton, was that it was far too slow for
the 15s time limit at the start of matches.
It was spending far too much time aligning itself to be perfect in angle
and so, due to that, took too long.
Charlie suggested that we should increase the window of precision in
order to speed it up.

Another thing noticed on the day is that Hudson removed some metal prongs
at the front of the robot that aided in it parking through catching the
parking goal and rotating the robot.
Due to this, the previous auton no longer worked and we had to resort
to making a simple park auton (that we didn't end up using).

Also, we decided to just use the old auton, as it's been proven to work
somewhat and just tweak the values to work on the comp fields.

Also, Hudson prototyped a pneumatics double-park system on the day that
required some special controls that I've implemented.
They didn't end up working, and he removed it.

Also some motors in the tower were the wrong port configs (changed).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>