Aquinus (state) comp commit
Changes made:
  * autonomous.zig:
    * (-) removed testing routine
    * (~) super jank left auton that scores 1 block sometimes
  * (~) drive.zig: updated drivetrain port
  * (-) odom.zig: removed redndant odom debug log
  * (+) pid.zig: added an experimental fast rotation func (doesn't work)

Saturday was our state competition at Aquinas college.
Our auton throughout the entire comp was HORRIBLE, just TERRIBLE.
But we still managed to get into grand finals (and therefore Adelaide)
only being defeated by our sister team 97230D and Delilah.

The main change to auton that I should focus on, is the way in which I
write it.
Each time I already have an auton from the last comp, and I keep tweaking
it over and over again to hope that a magic number pops up and it works
perfectly.
This is not the case.
Instead I should start over each time and plan, as our current auton is
ridiculously jank and just doesn't work.

Also, we should do a little bit, check if it's right, then continue.
That's the proper way to do it.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>